ipv6: Update link-local route when configuring network device
The IPv6 link-local address is generated during probing.
So if the link-local address of the hardware is changed
later on the IPv6 link-local address will become incorrect.
So we should set the link-local route for IPv6 before
before the network configuration is started; this way
we will always have a correct link-local address.

Signed-off-by: Hannes Reinecke <hare at suse.de>

tblume: modified rejected part