ibft: IPv6 support
The iBFT table is defined to be used with both, IPv4 and IPv6.
So update the iBFT handling to work with IPv6 addresses, too.
This patch updates the iBFT handling to extract the net_device
from the routing tables, which also removes the long-standing
hack with 'last_used_netdev'.

Signed-off-by: Hannes Reinecke <hare at suse.de>

tblume: modified rejected part